SIU/requirements.txt
_libgcc_mutex== 0.1
_openmp_mutex== 5.1
accelerate== 0.21.0
aiofiles== 23.2.1
aiohttp== 3.9.1
aiosignal== 1.3.1
altair== 5.2.0
annotated-types== 0.6.0
anyio== 3.7.1
appdirs== 1.4.4
async-timeout== 4.0.3
attrs== 23.1.0
bitsandbytes== 0.41.0
bzip2== 1.0.8
ca-certificates== 2023.08.22
certifi== 2023.11.17
charset-normalizer== 3.3.2
click== 8.1.7
cmake== 3.28.1
colorama== 0.4.6
contourpy== 1.2.0
cycler== 0.12.1
datasets== 2.18.0
deepspeed== 0.12.6
dill== 0.3.8
distro== 1.9.0
docker-pycreds== 0.4.0
easyjailbreak== 0.1.2
einops== 0.6.1
einops-exts== 0.0.4
et-xmlfile== 1.1.0
exceptiongroup== 1.2.0
fastapi== 0.105.0
ffmpy== 0.3.1
filelock== 3.13.1
flash-attn== 2.5.6
fonttools== 4.46.0
frozenlist== 1.4.1
fschat== 0.2.36
fsspec== 2023.12.2
gitdb== 4.0.11
gitpython== 3.1.40
gradio== 4.16.0
gradio-client== 0.8.1
h11== 0.14.0
hjson== 3.1.0
httpcore== 0.17.3
httpx== 0.24.0
huggingface-hub== 0.19.4
idna== 3.6
importlib-resources== 6.1.2
jinja2== 3.1.2
joblib== 1.3.2
jsonlines== 4.0.0
jsonschema== 4.20.0
jsonschema-specifications== 2023.11.2
kiwisolver== 1.4.5
ld_impl_linux-64== 2.38
libffi== 3.4.4
libgcc-ng== 11.2.0
libgomp== 11.2.0
libstdcxx-ng== 11.2.0
libuuid== 1.41.5
linkify-it-py== 2.0.2
lit== 17.0.6
llava== 1.2.2.post1
markdown-it-py== 2.2.0
markdown2== 2.4.12
markupsafe== 2.1.3
matplotlib== 3.8.2
mdit-py-plugins== 0.3.3
mdurl== 0.1.2
mpmath== 1.3.0
multidict== 6.0.4
multiprocess== 0.70.16
ncurses== 6.4
networkx== 3.2.1
nh3== 0.2.17
ninja== 1.11.1.1
nltk== 3.8.1
numpy== 1.26.2
nvidia-cublas-cu11== 11.10.3.66
nvidia-cublas-cu12== 12.1.3.1
nvidia-cuda-cupti-cu11== 11.7.101
nvidia-cuda-cupti-cu12== 12.1.105
nvidia-cuda-nvrtc-cu11== 11.7.99
nvidia-cuda-nvrtc-cu12== 12.1.105
nvidia-cuda-runtime-cu11== 11.7.99
nvidia-cuda-runtime-cu12== 12.1.105
nvidia-cudnn-cu11== 8.5.0.96
nvidia-cudnn-cu12== 8.9.2.26
nvidia-cufft-cu11== 10.9.0.58
nvidia-cufft-cu12== 11.0.2.54
nvidia-curand-cu11== 10.2.10.91
nvidia-curand-cu12== 10.3.2.106
nvidia-cusolver-cu11== 11.4.0.1
nvidia-cusolver-cu12== 11.4.5.107
nvidia-cusparse-cu11== 11.7.4.91
nvidia-cusparse-cu12== 12.1.0.106
nvidia-nccl-cu11== 2.14.3
nvidia-nccl-cu12== 2.18.1
nvidia-nvjitlink-cu12== 12.4.99
nvidia-nvtx-cu11== 11.7.91
nvidia-nvtx-cu12== 12.1.105
openai== 1.13.3
openpyxl== 3.1.2
openssl== 3.0.12
orjson== 3.9.10
packaging== 23.2
pandas== 2.1.4
peft== 0.4.0
pillow== 10.1.0
pip== 24.0
prompt-toolkit== 3.0.43
protobuf== 4.25.1
psutil== 5.9.7
py-cpuinfo== 9.0.0
pyarrow== 15.0.2
pyarrow-hotfix== 0.6
pydantic== 2.6.3
pydantic-core== 2.16.3
pydub== 0.25.1
pygments== 2.17.2
pynvml== 11.5.0
pyparsing== 3.1.1
python== 3.10.13
python-dateutil== 2.8.2
python-multipart== 0.0.6
pytz== 2023.3.post1
pyyaml== 6.0.1
readline== 8.2
referencing== 0.32.0
regex== 2023.10.3
requests== 2.31.0
rich== 13.7.1
rouge== 1.0.1
rpds-py== 0.15.2
ruff== 0.3.0
safetensors== 0.4.1
scikit-learn== 1.2.2
scipy== 1.11.4
semantic-version== 2.10.0
sentencepiece== 0.1.99
sentry-sdk== 1.39.1
setproctitle== 1.3.3
setuptools== 68.2.2
shellingham== 1.5.4
shortuuid== 1.0.11
six== 1.16.0
smmap== 5.0.1
sniffio== 1.3.0
sqlite== 3.41.2
starlette== 0.27.0
svgwrite== 1.4.3
sympy== 1.12
threadpoolctl== 3.2.0
tiktoken== 0.6.0
timm== 0.6.13
tk== 8.6.12
tokenizers== 0.15.1
tomlkit== 0.12.0
toolz== 0.12.0
torch== 2.1.2
torchvision== 0.16.2
tqdm== 4.66.1
transformers== 4.37.2
transformers-stream-generator== 0.0.5
triton== 2.1.0
typer== 0.9.0
typing-extensions== 4.9.0
tzdata== 2023.3
uc-micro-py== 1.0.2
urllib3== 2.1.0
uvicorn== 0.24.0.post1
wandb== 0.16.1
wavedrom== 2.0.3.post3
wcwidth== 0.2.13
websockets== 11.0.3
wheel== 0.41.2
xxhash== 3.4.1
xz== 5.4.5
yarl== 1.9.4
zlib== 1.2.13
